Improved exclusion mechanism
