+++ src/lib.rs
@@ -10,11 +10,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! bevy_seedling = "0.2"
-//!
-//! # At this stage, it may be better to track the main branch
-//! [dependencies]
-//! bevy_seedling = { git = "https://github.com/corvusprudens/bevy_seedling.git" }
+//! bevy_seedling = "0.3"
 //! ```
 //!
 //! Then, you'll need to add the [`SeedlingPlugin`] to your app.
